fixbingo

+++ src/migrations/1719990000000-AddAvatarDataToUsersTable.ts
@@ -2,10 +2,16 @@ import { MigrationInterface, QueryRunner } from 'typeorm';
 
 export class AddAvatarDataToUsersTable1719990000000 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(`ALTER TABLE "users" ADD COLUMN "avatarData" jsonb`);
+    const hasColumn = await queryRunner.hasColumn('users', 'avatarData');
+    if (!hasColumn) {
+      await queryRunner.query(`ALTER TABLE "users" ADD COLUMN "avatarData" jsonb`);
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(`ALTER TABLE "users" DROP COLUMN "avatarData"`);
+    const hasColumn = await queryRunner.hasColumn('users', 'avatarData');
+    if (hasColumn) {
+      await queryRunner.query(`ALTER TABLE "users" DROP COLUMN "avatarData"`);
+    }
   }
 }
